changes following review by samuelcolvin 2022-08-08.  Will rebase next
